refactor(provider): extract completion.rs tests into completion_tests.rs
completion.rs was already ~1705 lines; the two new built-in provider
golden-table rows tipped the Server Size Guard (MAX_LINES=1500). Move the
1149-line test module to a sibling completion_tests.rs attached via
#[path], keeping it in src/ so its include_str! references (completion.rs
plus cross-crate ../../djinn-agent and ../../djinn-control-plane sources)
resolve unchanged. completion.rs is now 570 lines.
